+++ pkgs/bazel_worker/e2e_test/bin/async_worker.dart
@@ -9,6 +9,6 @@ import 'package:e2e_test/async_worker.dart';
 
 /// This worker can run in one of two ways: normally, using stdin/stdout, or
 /// in an isolate, communicating over a [SendPort].
-Future main(List<String> args, [SendPort sendPort]) async {
+Future main(List<String> args, [SendPort? sendPort]) async {
   await ExampleAsyncWorker(sendPort).run();
 }